+++ util/service/registry.go
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+// 包含服务注册、服务发现相关的定义、实现
 package rocserv
 
 import (
