+++ util/service/registry_test.go
